requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
azure-core==1.32.0
azure-search-documents==11.5.2
azure-identity==1.16.0
python-dotenv==1.0.1
gunicorn==23.0.0
typing-extensions==4.12.2
pydantic==2.10.6
starlette==0.46.1
openai==1.68.2
tiktoken==0.9.0
numpy==2.2.4
requests==2.31.0
python-multipart==0.0.9